+++ bindings/python/expose-manifold.cpp
@@ -4,11 +4,15 @@
 #include "proxsuite-nlp/modelling/spaces/multibody.hpp"
 #endif
 
-#include "proxsuite-nlp/python/fwd.hpp"
+#include "proxsuite-nlp/python/polymorphic.hpp"
 #include "proxsuite-nlp/manifold-base.hpp"
 #include "proxsuite-nlp/modelling/spaces/cartesian-product.hpp"
 #include "proxsuite-nlp/modelling/spaces/vector-space.hpp"
 #include "proxsuite-nlp/modelling/spaces/tangent-bundle.hpp"
+
+#include <eigenpy/std-vector.hpp>
+
+#include <vector>
 
 namespace proxsuite {
 namespace nlp {
@@ -18,12 +22,13 @@ using context::Manifold;
 using context::MatrixRef;
 using context::Scalar;
 using context::VectorRef;
-using ManifoldPtr = shared_ptr<Manifold>;
+using PolyManifold = polymorphic<Manifold>;
 using CartesianProduct = CartesianProductTpl<Scalar>;
 
 void exposeManifoldBase() {
   using context::MatrixXs;
   using context::VectorXs;
+  register_polymorphic_to_python<PolyManifold>();
 
   using BinaryFunTypeRet = VectorXs (Manifold::*)(const ConstVectorRef &,
                                                   const ConstVectorRef &) const;
@@ -32,41 +37,41 @@ void exposeManifoldBase() {
   using JacobianFunType = void (Manifold::*)(
       const ConstVectorRef &, const ConstVectorRef &, MatrixRef, int) const;
 
-  bp::class_<Manifold, ManifoldPtr, boost::noncopyable>(
+  bp::class_<Manifold, boost::noncopyable>(
       "ManifoldAbstract", "Manifold abstract class.", bp::no_init)
       .add_property("nx", &Manifold::nx, "Manifold representation dimension.")
       .add_property("ndx", &Manifold::ndx, "Tangent space dimension.")
-      .def("neutral", &Manifold::neutral, bp::args("self"),
+      .def("neutral", &Manifold::neutral, "self"_a,
            "Get the neutral point from the manifold (if a Lie group).")
-      .def("rand", &Manifold::rand, bp::args("self"),
+      .def("rand", &Manifold::rand, "self"_a,
            "Sample a random point from the manifold.")
-      .def("isNormalized", &Manifold::isNormalized, bp::args("self", "x"),
+      .def("isNormalized", &Manifold::isNormalized, ("self"_a, "x"),
            "Check if the input vector :math:`x` is a viable element of the "
            "manifold.")
       .def<BinaryFunType>("integrate", &Manifold::integrate,
-                          bp::args("self", "x", "v", "out"))
+                          ("self"_a, "x", "v", "out"))
       .def<BinaryFunType>("difference", &Manifold::difference,
-                          bp::args("self", "x0", "x1", "out"))
+                          ("self"_a, "x0", "x1", "out"))
       .def<BinaryFunTypeRet>("integrate", &Manifold::integrate,
-                             bp::args("self", "x", "v"))
+                             ("self"_a, "x", "v"))
       .def<BinaryFunTypeRet>("difference", &Manifold::difference,
-                             bp::args("self", "x0", "x1"))
+                             ("self"_a, "x0", "x1"))
       .def("interpolate",
            (void(Manifold::*)(const ConstVectorRef &, const ConstVectorRef &,
                               const Scalar &, VectorRef)
                 const)(&Manifold::interpolate),
-           bp::args("self", "x0", "x1", "u", "out"))
+           ("self"_a, "x0", "x1", "u", "out"))
       .def(
           "interpolate",
           (VectorXs(Manifold::*)(const ConstVectorRef &, const ConstVectorRef &,
                                  const Scalar &) const)(&Manifold::interpolate),
-          bp::args("self", "x0", "x1", "u"),
+          ("self"_a, "x0", "x1", "u"),
           "Interpolate between two points on the manifold. Allocated version.")
       .def<JacobianFunType>("Jintegrate", &Manifold::Jintegrate,
-                            bp::args("self", "x", "v", "Jout", "arg"),
+                            ("self"_a, "x", "v", "Jout", "arg"),
                             "Compute the Jacobian of the exp operator.")
       .def<JacobianFunType>("Jdifference", &Manifold::Jdifference,
-                            bp::args("self", "x0", "x1", "Jout", "arg"),
+                            ("self"_a, "x0", "x1", "Jout", "arg"),
                             "Compute the Jacobian of the log operator.")
       .def(
           "Jintegrate",
@@ -76,10 +81,10 @@ void exposeManifoldBase() {
             m.Jintegrate(x, v, Jout, arg);
             return Jout;
           },
-          bp::args("self", "x", "v", "arg"),
+          ("self"_a, "x", "v", "arg"),
           "Compute and return the Jacobian of the exp.")
       .def("JintegrateTransport", &Manifold::JintegrateTransport,
-           bp::args("self", "x", "v", "J", "arg"),
+           ("self"_a, "x", "v", "J", "arg"),
            "Perform parallel transport of matrix J expressed at point x+v to "
            "point x.")
       .def(
@@ -90,20 +95,25 @@ void exposeManifoldBase() {
             m.Jdifference(x0, x1, Jout, arg);
             return Jout;
           },
-          bp::args("self", "x", "v", "arg"),
+          ("self"_a, "x0", "x1", "arg"),
           "Compute and return the Jacobian of the log.")
       .def("tangent_space", &Manifold::tangentSpace, bp::args("self"),
            "Returns an object representing the tangent space to this manifold.")
       .def(
           "__mul__",
-          +[](const ManifoldPtr &a, const ManifoldPtr &b) { return a * b; })
+          +[](const PolyManifold &a, const PolyManifold &b) { return a * b; })
       .def(
-          "__mul__", +[](const ManifoldPtr &a,
+          "__mul__", +[](const PolyManifold &a,
                          const CartesianProduct &b) { return a * b; })
       .def(
-          "__rmul__", +[](const ManifoldPtr &a, const CartesianProduct &b) {
+          "__rmul__", +[](const PolyManifold &a, const CartesianProduct &b) {
             return a * b;
           });
+
+  eigenpy::StdVectorPythonVisitor<std::vector<PolyManifold>>::expose(
+      "StdVec_Manifold",
+      eigenpy::details::overload_base_get_item_for_std_vector<
+          std::vector<PolyManifold>>());
 }
 
 /// Expose the tangent bundle of a manifold type @p M.
@@ -112,13 +122,12 @@ bp::class_<TangentBundleTpl<M>, bp::bases<Manifold>>
 exposeTangentBundle(const char *name, const char *docstring) {
   using OutType = TangentBundleTpl<M>;
   return bp::class_<OutType, bp::bases<Manifold>>(
-             name, docstring, bp::init<M>(bp::args("self", "base")))
-      .add_property(
-          "base",
-          bp::make_function(
-              &OutType::getBaseSpace,
-              bp::return_value_policy<bp::reference_existing_object>()),
-          "Get the base space.");
+             name, docstring, bp::init<M>(("self"_a, "base")))
+      .add_property("base",
+                    bp::make_function(&OutType::getBaseSpace,
+                                      bp::return_internal_reference<>()),
+                    "Get the base space.")
+      .def(PolymorphicVisitor<PolyManifold>());
 }
 
 /// Expose the tangent bundle with an additional constructor.
@@ -135,8 +144,10 @@ void exposeCartesianProduct();
 /// and no-arg default constructor.
 template <typename LieGroup>
 void exposeLieGroup(const char *name, const char *docstring) {
-  bp::class_<PinocchioLieGroup<LieGroup>, bp::bases<Manifold>>(
-      name, docstring, bp::init<>(bp::args("self")));
+  bp::class_<PinocchioLieGroup<LieGroup>, bp::bases<Manifold>> cl(
+      name, docstring, bp::init<>("self"_a));
+  cl.def(PolymorphicVisitor<PolyManifold>());
+  cl.enable_pickling_(true);
 }
 
 void exposePinocchioSpaces() {
@@ -150,7 +161,8 @@ void exposePinocchioSpaces() {
   bp::class_<PinocchioLieGroup<DynSizeEuclideanSpace>, bp::bases<Manifold>>(
       "EuclideanSpace", "Pinocchio's n-dimensional Euclidean vector space.",
       bp::no_init)
-      .def(bp::init<int>(bp::args("self", "dim")));
+      .def(bp::init<int>(("self"_a, "dim")))
+      .def(PolymorphicVisitor<PolyManifold>());
 
   exposeLieGroup<VectorSpaceOperationTpl<1, Scalar>>(
       "R", "One-dimensional Euclidean space AKA real number line.");
@@ -194,13 +206,14 @@ void exposePinocchioSpaces() {
       .add_property("model",
                     bp::make_function(&Multibody::getModel,
                                       bp::return_internal_reference<>()),
-                    "Return the Pinocchio model instance.");
+                    "Return the Pinocchio model instance.")
+      .def(PolymorphicVisitor<PolyManifold>());
 
   using MultiPhase = MultibodyPhaseSpace<Scalar>;
   bp::class_<MultiPhase, bp::bases<Manifold>>(
       "MultibodyPhaseSpace",
       "Tangent space of the multibody configuration group.",
-      bp::init<const Model &>(bp::args("self", "model")))
+      bp::init<const Model &>(("self"_a, "model")))
       .add_property("model",
                     bp::make_function(&MultiPhase::getModel,
                                       bp::return_internal_reference<>()),
@@ -209,7 +222,8 @@ void exposePinocchioSpaces() {
                                 +[](const MultiPhase &m) -> const Multibody & {
                                   return m.getBaseSpace();
                                 },
-                                bp::return_internal_reference<>()));
+                                bp::return_internal_reference<>()))
+      .def(PolymorphicVisitor<PolyManifold>());
 }
 #endif
 
@@ -220,7 +234,8 @@ void exposeManifolds() {
   /* Basic vector space */
   bp::class_<VectorSpaceTpl<Scalar>, bp::bases<Manifold>>(
       "VectorSpace", "Basic Euclidean vector space.", bp::no_init)
-      .def(bp::init<const int>(bp::args("self", "dim")));
+      .def(bp::init<const int>(("self"_a, "dim")))
+      .def(PolymorphicVisitor<PolyManifold>());
 
   exposeCartesianProduct();
 #ifdef PROXSUITE_NLP_WITH_PINOCCHIO