+++ bindings/python/expose-residual.cpp
@@ -29,15 +29,15 @@ void exposeResiduals() {
 
   expose_function<ManifoldDifferenceToPoint<Scalar>>(
       "ManifoldDifferenceToPoint", "Difference vector x (-) x0.",
-      bp::init<const shared_ptr<Manifold> &, const ConstVectorRef &>(
-          bp::args("self", "space", "target")))
+      bp::init<const polymorphic<Manifold> &, const ConstVectorRef &>(
+          ("self"_a, "space", "target")))
       .add_property("target", &ManifoldDifferenceToPoint<Scalar>::target_);
 
   expose_function<LinearFunctionDifferenceToPoint<Scalar>>(
       "LinearFunctionDifferenceToPoint",
       "Linear function of the vector difference to a reference point.",
-      bp::init<const shared_ptr<Manifold> &, VectorXs, MatrixXs, VectorXs>(
-          bp::args("self", "space", "target", "A", "b")));
+      bp::init<const polymorphic<Manifold> &, VectorXs, MatrixXs, VectorXs>(
+          ("self"_a, "space", "target", "A", "b")));
 }
 
 } // namespace python