+++ include/proxsuite-nlp/modelling/residuals/linear.hpp
@@ -3,6 +3,7 @@
 #include "proxsuite-nlp/function-base.hpp"
 #include "proxsuite-nlp/function-ops.hpp"
 #include "proxsuite-nlp/modelling/residuals/state-residual.hpp"
+#include "proxsuite-nlp/third-party/polymorphic_cxx14.hpp"
 
 namespace proxsuite {
 namespace nlp {
@@ -44,7 +45,7 @@ struct LinearFunctionDifferenceToPoint : ComposeFunctionTpl<_Scalar> {
 
   using Manifold = ManifoldAbstractTpl<Scalar>;
 
-  LinearFunctionDifferenceToPoint(const shared_ptr<Manifold> &space,
+  LinearFunctionDifferenceToPoint(const polymorphic<Manifold> &space,
                                   const ConstVectorRef &target,
                                   const ConstMatrixRef &A,
                                   const ConstVectorRef &b)