+++ include/proxsuite-nlp/modelling/residuals/state-residual.hpp
@@ -6,6 +6,7 @@
 
 #include "proxsuite-nlp/function-base.hpp"
 #include "proxsuite-nlp/manifold-base.hpp"
+#include "proxsuite-nlp/third-party/polymorphic_cxx14.hpp"
 
 namespace proxsuite {
 namespace nlp {
@@ -26,10 +27,10 @@ public:
   using Manifold = ManifoldAbstractTpl<Scalar>;
 
   /// Target point on the space.
-  typename Manifold::PointType target_;
-  shared_ptr<Manifold> space_;
+  VectorXs target_;
+  polymorphic<Manifold> space_;
 
-  ManifoldDifferenceToPoint(const shared_ptr<Manifold> &space,
+  ManifoldDifferenceToPoint(const polymorphic<Manifold> &space,
                             const ConstVectorRef &target)
       : Base(space->nx(), space->ndx(), space->ndx()), target_(target),
         space_(space) {